requirements.txt
# Web framework
streamlit

# Web scraping
requests
beautifulsoup4
playwright
crawl4ai

# Async support
asyncpg
aiohttp
asyncio
# Env + config
python-dotenv

# Timezone
pytz
tzdata  # dùng nếu chạy trên Alpine hoặc server không có zoneinfo

# Data handling
pandas

# Utility
glob2  # nếu anh dùng glob nâng cao

# Optional (nếu dùng các tính năng nâng cao của crawl4ai)
lxml
nest_asyncio